+++ server/migrations/018_workspace_issue_prefix.down.sql
@@ -1,0 +1,1 @@
+DROP FUNCTION IF EXISTS berry_issue_identifier(uuid, integer);